refactor(NumberConverter): Add logics to convert number - Replace interface with class in number converter module
