--- src/main/java/lotto/util/DefaultNumberConverter.java
@@ -1,20 +1,0 @@
-package lotto.util;
-
-import lotto.exception.GeneralExceptionMessages;
-
-public class DefaultNumberConverter implements NumberConverter{
-
-    @Override
-    public int convertNumber(String input) {
-        String trimmedInput = input.trim();
-        if (trimmedInput.isEmpty() || !trimmedInput.matches("\\d+")) {
-            throw new IllegalArgumentException(GeneralExceptionMessages.INVALID_NUMBER);
-        }
-
-        try {
-            return Integer.parseInt(trimmedInput);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(GeneralExceptionMessages.INVALID_NUMBER);
-        }
-    }
-}
